apps/liaison_brain/pyproject.toml
[project]
name = "coda-liaison-brain"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "psycopg-binary>=3.2.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "langgraph-checkpoint-postgres>=3.0.0",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-deepseek>=1.0.1",
    "langchain-anthropic>=1.3.1",
    "langchain-openai>=1.1.7",
    "sse-starlette>=2.0.0",
]
